refactor: split page.tsx into hook and focused components
Extract all state and event handlers into hooks/useBRIGController.ts.
Split page.tsx UI into ReferenceInput, RingsPanel, ControlPanel,
StatisticsPanel, BugReportModal, NavBar and ImagePropertiesPanel
components. page.tsx is now 122 lines and acts as a thin orchestrator.
